[Zellij] Trying a default layout
nvim on right, two splits on left. I think I want this to actually be
a predefined session instead of a default template. Opening this
template on every new tab is kinda annoying at the moment

Also want to try stacking the splits on the left, auto-starting
bin/rails, and shrinking bin/rails out of view